cleanup rule applicability test
