+++ server/src/graql/reasoner/atom/Atom.java
@@ -215,22 +215,24 @@ public abstract class Atom extends AtomicBase {
     /**
      * @return set of potentially applicable rules - does shallow (fast) check for applicability
      */
-    public Stream<Rule> getPotentialRules() {
+    public Stream<Rule> getPotentialRules(boolean useRuleCache) {
         boolean isDirect = getPattern().getProperties(IsaProperty.class).findFirst()
                 .map(IsaProperty::isExplicit).orElse(false);
 
         return getPossibleTypes().stream()
-                .flatMap(type -> tx().ruleCache().getRulesWithType(type, isDirect))
+                .flatMap(type -> tx().ruleCache().getRulesWithType(type, isDirect, useRuleCache))
                 .distinct();
     }
+
+    public Stream<Rule> getPotentialRules(){ return getPotentialRules(true);}
 
     /**
      * @return set of applicable rules - does detailed (slow) check for applicability
      */
-    public Stream<InferenceRule> getApplicableRules() {
+    public Stream<InferenceRule> getApplicableRules(boolean useRuleCache) {
         if (applicableRules == null) {
             applicableRules = new HashSet<>();
-            getPotentialRules()
+            getPotentialRules(useRuleCache)
                     .map(rule -> tx().ruleCache().getRule(rule, () -> new InferenceRule(rule, tx())))
                     .filter(this::isRuleApplicable)
                     .map(r -> r.rewrite(this))
@@ -238,6 +240,8 @@ public abstract class Atom extends AtomicBase {
         }
         return applicableRules.stream();
     }
+
+    public Stream<InferenceRule> getApplicableRules() { return getApplicableRules(true);}
 
     /**
      * @return true if the atom requires materialisation in order to be referenced
